fix(kernel): gate JWT M2M public surface behind kernel build tag
Addresses PR #447 review feedback:

- Gate the public JWT private-key M2M surface behind
  `cgo && databricks_kernel`. Move WithJWTPrivateKeyM2M and
  JWTPrivateKeyM2MConfig out of connector.go into a new tagged
  connector_kernel_jwt.go, and tag the jwtm2m package itself. The
  pure-Go Thrift path has no JWT-signing implementation, so a
  default (Thrift-only) build no longer exposes an auth option it
  cannot honor (Eric's divergence concern). The internal
  JWTM2MCredentialsProvider interface stays untagged so resolveKernelAuth
  still compiles in the default build.
- Skip telemetry on the kernel JWT M2M path: its Authenticate always
  returns a kernel-only error, so telemetry/feature-flag HTTP calls
  would fail every request and burn round-trips at connect. Extend the
  existing U2M telemetry-skip to cover it (bot F2).
- Fix the godoc option list: restore the contiguous bullet style (a
  blank line between items terminates the godoc list) and add the JWT
  bullet inline (bot F1).
- Update the stale supported-auth enumerations: resolveKernelAuth doc,
  validateKernelConfig docstring, and the default-case rejection
  message now include JWT private-key M2M (bot F2, first review).
