+++ auth/oauth/jwtm2m/jwtm2m_test.go
@@ -1,3 +1,5 @@
+//go:build cgo && databricks_kernel
+
 package jwtm2m
 
 import (
